Move Schema into type.js to break circluar dependency
