+++ src/runtime/tests/entity-test.ts
@@ -9,10 +9,9 @@
  */
 import {assert} from '../../platform/chai-web.js';
 import {Manifest} from '../manifest.js';
-import {Schema} from '../schema.js';
 import {Entity, EntityClass} from '../entity.js';
 import {IdGenerator, Id} from '../id.js';
-import {EntityType} from '../type.js';
+import {Schema, EntityType} from '../type.js';
 import {SYMBOL_INTERNALS} from '../symbols.js';
 
 describe('Entity', () => {
