+++ src/runtime/tests/particle-api-test.ts
@@ -17,8 +17,7 @@ import {Arc} from '../arc.js';
 import {Description} from '../description.js';
 import {IdGenerator} from '../id.js';
 import {Manifest} from '../manifest.js';
-import {Schema} from '../schema.js';
-import {EntityType} from '../type.js';
+import {Schema, EntityType} from '../type.js';
 import {Runtime} from '../runtime.js';
 import {SingletonStore} from '../store.js';
 import {Speculator} from '../../planning/speculator.js';
