+++ src/runtime/type-variable-info.ts
@@ -10,8 +10,7 @@
 
 import {assert} from '../platform/assert-web.js';
 
-import {Schema} from './schema.js';
-import {EntityType, SlotType, Type, TypeVariable, TypeLiteral} from './type.js';
+import {Schema, EntityType, SlotType, Type, TypeVariable, TypeLiteral} from './type.js';
 
 interface TypeVariableInfoLiteral {
   name: string;
