+++ webpack.config.js
@@ -14,7 +14,7 @@ const CircularDependencyPlugin = require('circular-dependency-plugin');
 const lib = './shells/lib';
 
 // Decrease MAX_CYCLES every time you eliminate circular dependencies from the codebase.
-const MAX_CYCLES = 15;
+const MAX_CYCLES = 14;
 let numCyclesDetected = 0;
 
 module.exports = {
